2 Actions added
Agents can now pick up items and put them inside of chests.

They are general actions => they consider all items and chests in existance.

Fixed a bug where the planner was using references instead of coppies of world state, which resulted in item duplication in the planning phaze, which resulted in strange behavior.